add helper for monitored thread pools
Adds a helper class for creating an ExecutionContext based
on a java thread pool. This includes using a factory with
named threads and setting up the spectator monitoring.
