--- atlas-lwcapi/src/main/scala/com/netflix/atlas/lwcapi/NamedThreadFactory.scala
@@ -1,27 +1,0 @@
-/*
- * Copyright 2014-2023 Netflix, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.netflix.atlas.lwcapi
-
-import java.util.concurrent.ThreadFactory
-
-class NamedThreadFactory(name: String) extends ThreadFactory {
-
-  def newThread(r: Runnable): Thread = {
-    val thread = new Thread(r, name)
-    thread.setDaemon(true)
-    thread
-  }
-}
